core/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0.1
django-cors-headers==4.9.0
dj-database-url==2.3.0
djangorestframework==3.16.1
idna==3.11
PyJWT==2.10.1
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
supabase==2.3.0
psycopg2-binary==2.9.9
daphne
channels
drf-spectacular==0.27.1
razorpay
Pillow



